microsservices/notification/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
pika==1.3.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10